feat: jumbotron createdAt long으로 변환

+++ data/src/main/java/doro/android/data/dto/JumbotronResponse.kt
@@ -14,7 +14,7 @@ data class JumbotronResponse(
     @SerializedName("thumbnailUrl")
     val thumbnailUrl: String? = null,
     @SerializedName("createdAt")
-    val createdAt: Int,
+    val createdAt: Long,
 ): Parcelable {
 
     fun toDomain(): Jumbotron {
