+++ domain/src/main/java/doro/android/domain/entity/Jumbotron.kt
@@ -4,5 +4,5 @@ data class Jumbotron(
     val title: String,
     val body: String,
     val thumbnailUrl: String,
-    val createdAt: Int,
+    val createdAt: Long,
 )
